Typescriptifying gcp.cloudfunctions (#3060)
