+++ src/deploy/functions/deploy.ts
@@ -38,7 +38,7 @@ export async function deploy(context: any, options: any, payload: any): Promise<
     );
 
     const source = await prepareFunctionsUpload(context, options);
-    context.existingFunctions = await gcp.cloudfunctions.listAll(context.projectId);
+    context.existingFunctions = await gcp.cloudfunctions.listAllFunctions(context.projectId);
     payload.functions = {
       triggers: options.config.get("functions.triggers"),
     };
